Add defaults package
This change adds a defaults package. These defaults are meant to
customize the behavior of the system. They are not part of the
controllers package, because these defaults may be shared with an
admission or mutating webhook at some point.

The defaults package is placed under a pkg/ directory, a standard for
signaling that a package is a public API. This conveys that the defaults
are intended to be modified. It does not provide any guarantees about
backwards compatibility.