Improve Makefile and deployment process
This change improves the Makefile by adding targets for each of the
container images, renaming ambiguous targets, and adding environment
variables to customize builds. It also removes an unnecessary
dependency on the "test" target.

In addition, it improves the deployment process by adding introducing
a way to dynamically generate a config file. In particular, it:

 - Introduces the config/cmd/configure.go program. This program uses a
   template file and environment variables to produce a config file.

 - Adds the config/defaults_template.yaml file as the default
   configuration template.

 - Removes the old config/example_config.yaml file, since it is no
   longer applicable.

 - Adds a "configure" phony target to the Makefile. This target invokes
   the configuration program with the default configuration template. It
   creates a config/defaults.yaml file, which is used by the controller.