Report the status and error when defaults fail
Setting and using defaults on a load test may fail. For instance,
consider a load test that specifies an unknown language for a component.
If the test does explicitly set images, the controller will try to add
them using defaults. Since the language is unknown, there are no default
images to fallback on. In this case, the test component could not be
scheduled.

When this would occur, there was no update to the status of the test.
The test would not officially terminate, and the error was only visible
in the controller logs. This change updates the status of the load test
if the there is a problem with the defaults. It terminates it and
and includes the precise error in the test's status.